Make repositories list more compact
The old layout is quite wasteful when it comes to screen space,
especially when more than 5 or 10 repositories need to be shown
on the screen.

Using a compact table should allow for 2-3 times more repositories
to be shown in the same space. It's also a more consistent layout,
which makes the list easier to skim through.