Major refactoring of the codebase, specially in the directory scanning parts.
I have removed the directory management features from Gitter and ported them to GitList,
since they are application-specific logic. I was unhappy with the way directories and multiple
directories were handled, so i altered almost everything.

The tests are OK, but the coverage is not good. This means that i'll need som help from all
GitList users that are actively testing the development version with multiple directories.

I have removed the directory caching feature for now, it will be re-implemented soon.