Added the ability to categorize repositories by adding a "category" file in the .git folder
The category is read the same way as the repository description is read: By looking in the repository .git folder ( if not bare ) or if bare, in the repository folder. Had to do some modifications of how filtering works, and made the conclusion that the plugin that is used to filter, does not have the ability to filter elements on more than one level. So i made it possible to filter every category instead.